feat: Add user account deletion functionality and show remaining posts
- Implemented DeleteAccount component for user account deletion.
- Added ShowRemainingPosts component to display user's articles.
- Updated routing in App.jsx to include new components.
- Enhanced BelowNavbarContent to show trending posts based on views and likes.
- Modified BottomContent to handle email subscription with loading state.
- Refactored EditUserPage to include a link for account deletion.
- Updated UserProfilePage to fetch and display user articles.
- Improved SinglePost component to conditionally render follow/delete buttons based on user ownership.
- Introduced timeAgo utility for consistent time formatting across components.
- Removed unused font file and added a new font file.

+++ Backend/controllers/post.controller.js
@@ -43,12 +43,12 @@ module.exports.getSinglePost = async (req, res) => {
   }
 };
 
-module.exports.getSinglePostByUser = async (req, res) => {
+module.exports.getPostsByUser = async (req, res) => {
   try {
-    const { userId } = req.params;
+    const userId  = req.params.userId;
 
     const posts = await PostModel.find({ postedBy: userId })
-      .populate("postedBy", "fullname  profilePicture")
+      .populate("postedBy")
       .sort({ createdAt: -1 });
 
     if (!posts || posts.length === 0) {
@@ -76,18 +76,25 @@ module.exports.deletePost = async (req, res) => {
 
 module.exports.deleteAllPosts = async (req, res) => {
   try {
-    const posts = await PostModel.find({});
+    
+    const userId = req.user._id;
+
+    const posts = await PostModel.find({ postedBy: userId });
+
     if (posts.length === 0) {
       return res.status(404).json({ message: "No posts found to delete" });
     }
-    const deletedPosts = await PostModel.deleteMany({});
+
+    const deletedPosts = await PostModel.deleteMany({ postedBy: userId });
+
     if (deletedPosts.deletedCount === 0) {
       return res.status(404).json({ message: "No posts found to delete" });
     }
+
     return res.status(200).json({ message: "All posts deleted successfully" });
-  } catch (error) {
-    return res
-      .status(500)
-      .json({ message: "Server Error", error: error.message });
+  } 
+  catch (error) {
+
+    return res.status(500).json({ message: "Server Error", error: error.message });
   }
 };
